Fix licence text extraction in script
We request info from RPM in NAME:LICENCE format. Later on, we divided the line
into name and licence. For extracting the licence, we used the ${var##}
bashism, which removes the longest substring matching a pattern (said pattern
being /.*:/) from the beginning of var-value. This was a bug, since if the
licence text contained a colon, we would remove all the text up to that colon.
The fix is to use ${var#}, which works similarly, but removes the shortest
match. Since the package name is guaranteed not to contain a colon, removing
the shortest any-text-followed-by-colon substring should always remove
the package name and leave the licence description intact.