Detect duplicate package names and compare EVR
It is possible for a package to have multiple versions installed
at the same time. The two most common examples are:
a) "kernel" and associated packages - many distros keep at least 2
   of those, so you can easily boot up an older kernel, if the newest
   one causes some issues
b) multilib packages (i.e. having both an x86_64 and i686 build
   of the same library)

Until now, vrms-rpm ignored this, displaying all packages by name only.
This commit makes vrms-rpm ask for epoch-version-release.arch information
for each package, and when duplicate packages are detected, print this
info to eliminate ambiguity as to which package it is describing.