Get rid of count_members()
This commit makes licence_classify() utilise a ReBuffer to store the
list of child nodes before the top-level node is returned. Previously,
the count_members() function was called to calculate the number of
children and allocate a top-level node of appropriate size. This
approach, while it avoided the need for a buffer, had a singificant
flaw: if the two functions ended up disagreeing on how many child-nodes
are needed, this could lead to the code writing past allocated memory
and segfaulting. This patch trades somewhat higher memory usage for
improved safety.