Add classifier-spdx fuzz target for afl++
The new target is an executable (`fuzz-classifier-spdx`) that reads all
data available on stdin and passes it to both strict and lenient SPDX
licence classifiers.

Usage:
* Install afl++ (https://github.com/AFLplusplus/AFLplusplus)
* Build & start fuzzing with `make fuzz` (or `make fuzz-coverage` for
  fuzzing with coverage data, see below). This starts fuzzing based on
  all the input files from `test/fuzz/input/*`, and can be stopped with
  Ctrl+C.
* In case of any crashes, the `test/fuzz/output/default/crashes`
  directory will contain inputs that cause `fuzz-classifier-spdx` to
  crash. They can be used to reproduce issues.
* (optional) Use https://github.com/mrash/afl-cov to view code coverage
  of the fuzzing process:

```
afl-cov --overwrite -d ./test/fuzz/output \
        --coverage-cmd "cat AFL_FILE | $PWD/build/fuzz-classifier-spdx" \
        --code-dir ./
xdg-open ./test/fuzz/output/cov/web/index.html
```

I only tried this on Ubuntu, and had to provide my own implementation of
`rpmvercmp` (not included here) for it to compile because it apparently
is missing from the librpm I got from the repo. Please double check if
it works well for you.

Signed-off-by: Marcin Radomski <marcin@mradomski.pl>