Fix SPDX classifier eating parentheses sometimes
This bug could be observed when a licence string began with an opening
parenthesis, like when a compound licence used lowercase joiners.
Some examples:
a) "(Licence-A or Licence-B) and Licence-C"
b) "(Licence-A or Licence-B) and (Licence-C or Licence-D)"

In strict mode, lowercase joiners are ignored, leading the algorithm
to classify this entire string as a single licence. The code would
detect that the string opens with a parenthesis and take that as
a sign that the whole string is parenthesized (e.g. "(text")).
It would then proceed to remove the opening parenthesis,
leading to a mangled string. In example b), the end-of-string
closing parenthesis would also be removed.