Search for WITH operator in loose classifier
Until now, the "loose" classifier handled licensing exceptions
by matching the end of the licence string against a list of acceptable
suffixes, each of those containing a variation of the WITH operator -
like " WITH " or "-with-". This was rather inelegant.

This commit changes the behaviour of the classifier to search for the
presence of the WITH operator within the licence string. Should it be
found, the string is split into the licence name and exception name.
Then, the exception name can be checked against a (slightly shorter)
list of acceptable suffixes.