Do not remember JSON nesting types
The JSON API is based on callbacks to eliminate the risk of mismatching
the number of open/close calls. As such, there isn't really a need
to remember the nested types inside the JSON document - every place
that closes a nested entity (array/object) can hard-code the type,
the same way as it happens when opening an entity.