Hide output generation behind an interface
This commit introduces a "Printer" interface to the codebase, which
is used to separate output generation from iteration over the list of
packages (which is now done through an opaque iterator type). This
allows for better separation of concerns within the code and makes it
easier to introduce more output formats (should the need arise).