Avoid heap allocation for JSON documents
The struct representing internal state of a JSON document was allocated
on the heap. This made sense back when this struct contained a small
stack. Now that it holds only a pointer and some ints, it makes sense
to just keep it on the stack and avoid involving the memory allocator.