Fix unitialized memory read
Inside licences_read(), there are two initialization steps: opening the
licence list file, and allocating the LicenceData struct. If opening the
file fails, a goto statement is used to jump to "cleanup" code, which
frees both pointers.

Commit 269fd347f23b634fbba6fe0df44c5a86a733db78 changed the ordering
of code inside the function, so that the struct is allocated only after
the file is successfully opened. However, it omitted to NULL-initialize
the pointer, which meant that if opening the file failed, the cleanup
code would call free() on an unitialized pointer.